Initial commit of the project

+++ src/main/webapp/js/script.js
@@ -102,9 +102,9 @@ function liveSearchByHalfUsername() {
 							+ '<td width="15%">' + data.email + '</td>'
 							+ '<td width="10%">' + data.role + '</td>'
 							+ '<td width="5%">' + data.enabled + '</td>'
-							+ '<td width="10%"> <a id="setasadminuser" href="set_as_admin_user.html?username=' + data.username + '&role=' + data.role + '">' + setAsAdminUser + '</a> </td>'
-							+ '<td width="10%"> <a id="lockunlockaccount" href="lock_account_by_username.html?username=' + data.username + '&enabled=' + data.enabled + '">' + lockUnlockAccount + '</a> </td>'
-							+ '<td width="10%"> <a id="deleteaccount" href="delete_account_by_username.html?username=' + data.username + '">' + deleteAccount + '</a> </td>'
+							+ '<td width="10%"> <a onclick="confirmDialog()" id="setasadminuser' + data.id + '" href="set_as_admin_user.html?username=' + data.username + '&role=' + data.role + '">' + setAsAdminUser + '</a> </td>'
+							+ '<td width="10%"> <a onclick="confirmDialog()" id="lockunlockaccount' + data.id + '" href="lock_account_by_username.html?username=' + data.username + '&enabled=' + data.enabled + '">' + lockUnlockAccount + '</a> </td>'
+							+ '<td width="10%"> <a onclick="confirmDialog()" id="deleteaccount' + data.id + '" href="delete_account_by_username.html?username=' + data.username + '">' + deleteAccount + '</a> </td>'
 							+ '</tr>');
 				}
 			});
